Wip: Stats clean up
