+++ core/src/main/scala/boon/sbt/BoonRunner.scala
@@ -6,24 +6,12 @@ import sbt.testing.TaskDef
 
 import boon.printers.PrinterSetting
 import boon.printers.SimplePrinter
-import boon.model.SuiteResult
-import boon.model.TestResult
-import boon.model.Failed
-import boon.model.Passed
+import boon.model.stats.SuiteStats
+import boon.model.stats.StatusCount
+import boon.model.stats.AssertionCount
 
 import java.util.concurrent.atomic.AtomicReference
 
-trait TestStatusListener {
-  def suiteResult(result: SuiteResult): Unit
-  def suiteFailed(reason: String): Unit
-}
-
-final case class StatusCount(passed: Int, failed: Int)
-
-final case class AssertionCount(statusCount: StatusCount, notRun: Int)
-
-//This could be a monoid
-final case class SuiteStats(suites: StatusCount, tests: StatusCount, assertions: AssertionCount)
 
 
 final class BoonRunner(
@@ -34,45 +22,6 @@ final class BoonRunner(
 
   private val statsVecAtomic = new AtomicReference[Vector[SuiteStats]](Vector.empty[SuiteStats])
 
-  //Should this run Async?
-  private val statusListener = new TestStatusListener {
-
-    def suiteResult(result: SuiteResult): Unit = {
-      val stats = SuiteStats(suites = StatusCount(0, 0), tests = StatusCount(0, 0), assertions = AssertionCount(StatusCount(0, 0), 0))
-      val suiteCounts =
-        SuiteResult.suiteResultToPassable(result) match {
-          case Passed =>  stats.copy(suites = stats.suites.copy(passed = stats.suites.passed + 1))
-          case Failed => stats.copy(suites = stats.suites.copy(failed = stats.suites.failed + 1))
-        }
-
-        val testCounts =
-          result.testResults.map(TestResult.testResultToPassable).foldLeft(suiteCounts) {
-            case (acc, Passed) => acc.copy(tests = acc.tests.copy(passed = acc.tests.passed + 1))
-            case (acc, Failed) => acc.copy(tests = acc.tests.copy(failed = acc.tests.failed + 1))
-          }
-
-        val newStats =
-          result.testResults.map(TestResult.testResultToAssertionCount).foldLeft(testCounts) {
-            case (acc, (pass, fail, notRun)) =>
-              acc.copy(assertions =
-                        acc.assertions.copy(statusCount =
-                                              acc.assertions.statusCount.copy(passed = acc.assertions.statusCount.passed + pass,
-                                                                              failed = acc.assertions.statusCount.failed + fail),
-                                              notRun = acc.assertions.notRun + notRun))
-          }
-
-        val _ = statsVecAtomic.updateAndGet(newStats +: _)
-        ()
-      }
-
-    def suiteFailed(reason: String): Unit = {
-      val stats = SuiteStats(suites = StatusCount(0, 0), tests = StatusCount(0, 0), assertions = AssertionCount(StatusCount(0, 0), 0))
-      val newStats = stats.copy(suites = stats.suites.copy(failed =stats.suites.failed + 1))
-      val _ = statsVecAtomic.updateAndGet(newStats +: _)
-      ()
-    }
-  }
-
   //use default printer for now, change to use from args
   override def tasks(list: Array[TaskDef]): Array[Task] = {
     list.map(new BoonTask(
@@ -80,7 +29,7 @@ final class BoonRunner(
                   classLoader,
                   (so, c, print) =>
                     SimplePrinter(so, PrinterSetting.defaults(c), print),
-                  statusListener
+                  new BoonTestStatusListener(statsVecAtomic)
              )
     )
   }
