+++ core/src/main/scala/boon/sbt/Monoid.scala
@@ -1,0 +1,8 @@
+package boon.sbt
+
+trait Monoid[F] {
+
+  def mempty: F
+
+  def mappend(x: F, y: F): F
+}
